Split the tuner source file between interface and implemenation
